Add sample GitHub issue templates
These templates are mostly copied from OctoPrint's own templates, but are
adapted to suit a plugin repository. Most notably, the bug_report includes
instructions on how to enable debug logging for the plugin.

The feature_request is also similar to OctoPrint's but does not try to
steer the user toward plugins, because, well... :)